\name{SolMultinomClass-package}
\alias{SolMultinomClass-package}
\alias{SolMultinomClass}
\docType{package}
\title{
\packageTitle{SolMultinomClass}
}
\description{
\packageDescription{SolMultinomClass}
}
\author{
\packageAuthor{SolMultinomClass}

Maintainer: \packageMaintainer{SolMultinomClass}
}
\keyword{ package }
